ai-microservice/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
prophet
faiss-cpu
pydantic
